task 2 for week4 project: Notes App: Lazy State Initialization
Challenge: Try to figure out a way to display only the
first line of note.body as the note summary in the
sidebar.

Hint 1: note.body has "invisible" newline characters
in the text every time there's a new line shown. E.g.
the text in Note 1 is:
"# Note summary\n\nBeginning of the note"

Hint 2: See if you can split the string into an array
using the "\n" newline character as the divider