task 7 for week4 project 2: Tenzies: Hold dice part 2
Challenge: Update the `holdDice` function to flip
the `isHeld` property on the object in the array
that was clicked, based on the `id` prop passed
into the function.

Hint: as usual, there's > 1 way to accomplish this.
I'll be using `dice.map()` and checking for the `id`
of the die to determine which one to flip `isHeld` on,
but you can do whichever way makes the most sense to you.