task 7 for week4 project 2: Tenzies: Hold dice part 3
Challenge: Update the `rollDice` function to not just roll
all new dice, but instead to look through the existing dice
to NOT role any that are being `held`.

Hint: this will look relatively similiar to the `holdDice`
function below. When creating new dice, remember to use
`id: nanoid()` so any new dice have an `id` as well.